Enable strict flags and fix corresponding errors (#1)

+++ src/main.c
@@ -84,7 +84,7 @@ int main(int argc, char *argv[])
     int quit = 0;
 
     const Uint8 *keyboard_state = SDL_GetKeyboardState(NULL);
-    const int delay_ms = roundf(1000.0 / GAME_FPS);
+    const Uint32 delay_ms = (Uint32) roundf(1000.0f / GAME_FPS);
     SDL_Event e;
     while (!quit) {
         while (SDL_PollEvent(&e)) {
